Tools for moving squares in diagramless puzzles.
These options allow moving all squares on the grid or all letters
adjacent to a given letter (those which are connected by non-empty,
non-black squares). This can be quite helpful in the early stages of
solving a diagramless puzzle when only relative but not absolute grid
locations are determined.

To use these tools, select a square in the grid, select one of the
move options from the Tools -> Diagramless menu, and then select a
destination square for the initially selected square. All squares in
the requested set will be translated by this amount. If this migration
would result in a loss of information, either because some letters
will move past the grid boundaries, or because some letters which
aren't being moved will be overwritten, a prompt is shown to either
cancel the move altogether, decline the move but provide the ability
to select a different square, or perform the move anyway.

Inspired by the similar features in Diagnil.