Update wxWidgets dependency to 3.1.2.
The smooth scroll patch is no longer necessary as this fix was
included in 3.1.1. The other patches have been rebased to apply
cleanly.

A missing import has been added to PreferencesPanel.cpp to fix Windows
builds.

The focus-on-activate hack is no longer necessary (and does not build
on Mac, where m_winLastFocused is not available); this hack dates back
to 2010 in Xword and was fixed in wxWidgets 2.9.3 per
https://trac.wxwidgets.org/ticket/1599; a similar bug was fixed for
Mac (leading to the variable being removed) in 3.0.2 per
https://trac.wxwidgets.org/ticket/17149.

Checks have been added to wxColour to ensure that any attempt to get
an attribute (e.g. Red()/Green()/Blue() or OSXGetNSColor() on Mac)
will fail for wxNullColour, which was being used as a placeholder in
some places. Add guards to prevent this from happening.

HtmlCell::adjustPagebreaks no longer takes an argument with prior
pagebreaks. We can just remove it; we don't worry about pagebreaks
since we always render to fit content on a single page at a time.

An upstream patch has been added to fix a black background that
appears on the Mac toolbar for selected icons.

Many wxStatusBar changes due to Mac rendering changes/bugs:
-Work around https://trac.wxwidgets.org/ticket/18469 by disabling
 wxST_SIZEGRIP on Mac since it just results in invisible padding.
 OS X 10.7+ doesn't show size grips in general, so this is consistent.
-Switch to wxStaticText for all fields; the default text is colored
 differently when inactive and we can't easily replicate this, so this
 allows us to keep a consistent style with the timer.
-Vertically center align all the wxStaticText fields since the new
 status bar is taller than the text.
-Work around https://trac.wxwidgets.org/ticket/18470 with a wxWidgets
 patch avoiding a crash when setting the background color of the alert
 SizedText to null.

Fixes #38