Fix delay opening files on OS X Mojave.
Bug was identified and fixed upstream in
https://trac.wxwidgets.org/ticket/18305; backport the patch to
wxWidgets 3.1.0 and use it here. Also add wxWidgets version to the
build directory to avoid reusing a cached wxWidgets build without the
patch. In the future, if we add more patches, we can add another
identifier to the directory to identify the set of patches.

Fixes #67