Disable smart quotes in metadata settings text boxes.
Since these can be scripts, we don't want fancy quotes to be
automatically substituted.

This bug also impacts some of the xworddebug lua plugin UI, but this
would be harder to fix since the generated bindings don't include
WXOSX-specific methods, and this is less impactful since it's debug
only.

See #70