Fix md5 implementation inconsistency on Mac.
Fresh installs on Windows were attempting to regenerate Lua UI using
py2lua despite the MD5 sums matching up on Mac. This was because the
md5 module was returning incorrect results on Mac, which was because
wxLua replaces the 'bit' module built into LuaJIT with one that
behaves differently (returns unsigned integers instead of signed
ones). We should be able to disable the wxLua 'bit' replacement since
we are using LuaJIT on all platforms and since LuaJIT's implementation
is explicitly built to be consistent across platforms and
architectures.

See also pkulchenko/wxlua#43.

Fixes #61