Fix initial window positioning on Windows.
For some reason, on Windows with wxWidgets 3.1.3, a window at the left
edge of the screen has a small negative left coordinate instead of 0
as expected. Rather than snap negative positions to 20 pixels, we
allow slightly negative positions without repositioning them on the
next startup. We still reposition windows that are significantly
offscreen in case this was unintentional or the saved position was
invalid.

Fixes #84