Consolidate to one version code throughout the build.
Normal builds without special setup will read the version from VERSION,
which would generally be expected to be a SNAPSHOT version. Release
builds (those generated by CI with tags on GitHub) get the tag name as
the version. This version is propagated to the Windows installer, the
app, and the Mac plist file.

dist/gen_packages.lua still has a hard-coded reference; this will be
handled separately.

This commit also fixes a bug in xword.pkgmgr.is_newer, which seems to
have been passing an escaped "." character to pl.stringx.split when a
non-escaped character was expected. We also extend is_newer to support
-SNAPSHOT builds by considering them to be older than the non-SNAPSHOT
build of the same version, but newer than any previous version.

See #40