Complete configuration for auto updater.
- Generate packages_windows.lua and packages_mac.lua as part of release
builds and include them in the release artifacts. For now, since we are
only distributing plugins inside the distribution, we can use a simple
template and a sed-based find/replace to generate these files.
dist/gen_packages.lua works for this purpose as well, but requires Lua
to be installed in the CI environment, which is non-trivial. We can
always revisit if we decide to distribute plugin packages separately.

- Point the auto-updater to a known location on GitHub pages to search
for new packages.

With this along with #102 and #103 (along with #106 for Windows), the
full distribution workflow is functional on Windows/Mac. Pushing a new
build consists of the following steps:

1) Create a GitHub release. Set the tag to a valid release number, e.g.
"0.7.0".
2) Wait for CI builds to complete successfully on AppVeyor (Windows) and
Travis CI (Mac).
3) Validate that the builds are functioning as expected.
4) Download packages_windows.lua and packages_mac.lua from the builds to
docs/, commit, and push (to master). This will cause the new build to go
"live" and start showing up for auto updates.

See #40