Hacky workaround for segfault when closing rebus window on Linux.
We're seeing crashes when trying to close a dialog containing a
wxTextCtrl in the EVT_CHAR event handler for that control. It
seems that in the course of destroying the text control,
wxWidgets tries to call an API on it related to handling key
presses which fails because it is being disposed. To work around
this, defer the window destroy until after the keypress event by
defining a new custom event and sending it. (There is probably
a better way to do this, but I'm not familiar enough with wxWidgets
to know it).