Clean up distinction between clue list direction and display title.
- Use the direction (string key in Clues map) when we need a
  semantically-meaningful value.
- Use the title (ClueList#getTitle) when we need a value to be displayed
  to the user.
- Handle ipuz files that contain both a direction and display title. For
  other formats, the two values are identical.

In the UI, the clue list pane and printout have been updated to use the
display title.

When saving JPZ files, we use the display title; if we have to choose
one, this feels like the better choice, since we shouldn't need to rely
heavily on the semantic meaning when the words are explicitly given, and
the display title is what the user sees.

Note that the Lua scripts have not been updated, as the display title is
not yet exposed there, and this will be a slightly more involved change
(see #220).

See #219