Fix build-luajit.sh to point to the git repository.
Per https://luajit.org/download.html, release tarballs have been taken
down. For deterministic results, it seems best to point to a specific
tag in the Git repository, and for the moment we can just use the same
version we were using. In the future, we could consider following the
recommendation to point to a newer version or at least a fixed commit on
a newer branch.

We use the official GitHub mirror rather than the main repository in
order to support "--depth=1" to avoid fetching the entire repository
instead of just the branch we care about.